remove depency on c99 types
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@13211 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/roken/resolve.h
@@ -176,12 +176,12 @@ typedef int HEADER; /* will never be used */
 #endif
 
 struct dns_header {
-    u_int16_t id;
-    u_int16_t misc;
-    u_int16_t qdcount;
-    u_int16_t ancount;
-    u_int16_t nscount;
-    u_int16_t arcount;
+    unsigned id;
+    unsigned misc;
+    unsigned qdcount;
+    unsigned ancount;
+    unsigned nscount;
+    unsigned arcount;
 };
 
 struct dns_reply{
